refactor: globalSetup and create cookie manually
