+++ .eslintrc.yaml
@@ -35,13 +35,7 @@ rules:
   "@typescript-eslint/no-extra-semi": off
   eqeqeq: error
   import/order:
-    [
-      error,
-      {
-        alphabetize: { order: "asc" },
-        groups: [["builtin", "external", "internal"], "parent", "sibling"],
-      },
-    ]
+    [error, { alphabetize: { order: "asc" }, groups: [["builtin", "external", "internal"], "parent", "sibling"] }]
   no-async-promise-executor: off
   # This isn't a real module, just types, which apparently doesn't resolve.
   import/no-unresolved: [error, { ignore: ["express-serve-static-core"] }]
